Tidy up ContainerMock
