+++ CosmosTestHelpers/ContainerMockData/ContainerItem.cs
@@ -1,48 +1,47 @@
 ﻿using Microsoft.Azure.Cosmos;
 using Newtonsoft.Json;
 
-namespace CosmosTestHelpers.ContainerMockData
+namespace CosmosTestHelpers.ContainerMockData;
+
+internal class ContainerItem
 {
-    internal class ContainerItem
+    public ContainerItem(string id, string json, PartitionKey partitionKey, int? expiryTime)
     {
-        public ContainerItem(string id, string json, PartitionKey partitionKey, int? expiryTime)
-        {
-            Id = id;
-            ExpiryTime = expiryTime;
-            Json = json;
-            PartitionKey = partitionKey;
-            ETag = GenerateETag();
-        }
+        Id = id;
+        ExpiryTime = expiryTime;
+        Json = json;
+        PartitionKey = partitionKey;
+        ETag = GenerateETag();
+    }
 
-        public PartitionKey PartitionKey { get; }
-        public string Id { get; }
-        public string Json { get; }
-        public string ETag { get; private set; }
-        public int? ExpiryTime { get; }
+    public PartitionKey PartitionKey { get; }
+    public string Id { get; }
+    public string Json { get; }
+    public string ETag { get; private set; }
+    public int? ExpiryTime { get; }
 
-        public bool RequireETagOnNextUpdate { get; private set; }
-        public bool HasScheduledETagMismatch { get; private set; }
+    public bool RequireETagOnNextUpdate { get; private set; }
+    public bool HasScheduledETagMismatch { get; private set; }
 
-        public T Deserialize<T>()
-        {
-            return JsonConvert.DeserializeObject<T>(Json);
-        }
+    public T Deserialize<T>()
+    {
+        return JsonConvert.DeserializeObject<T>(Json);
+    }
 
-        public void ScheduleMismatchETagOnNextUpdate()
-        {
-            RequireETagOnNextUpdate = true;
-            HasScheduledETagMismatch = true;
-        }
+    public void ScheduleMismatchETagOnNextUpdate()
+    {
+        RequireETagOnNextUpdate = true;
+        HasScheduledETagMismatch = true;
+    }
 
-        public void ChangeETag()
-        {
-            ETag = GenerateETag();
-            HasScheduledETagMismatch = false;
-        }
+    public void ChangeETag()
+    {
+        ETag = GenerateETag();
+        HasScheduledETagMismatch = false;
+    }
         
-        private static string GenerateETag()
-        {
-            return $"\"{Guid.NewGuid()}\"";
-        }
+    private static string GenerateETag()
+    {
+        return $"\"{Guid.NewGuid()}\"";
     }
 }
