+++ CosmosTestHelpers/ContainerMockData/ContainerMockException.cs
@@ -2,39 +2,38 @@
 using System.Runtime.Serialization;
 using Microsoft.Azure.Cosmos;
 
-namespace CosmosTestHelpers.ContainerMockData
+namespace CosmosTestHelpers.ContainerMockData;
+
+[Serializable]
+public class ContainerMockException : Exception
 {
-    [Serializable]
-    public class ContainerMockException : Exception
+    public HttpStatusCode StatusCode { get; }
+
+    public ContainerMockException(HttpStatusCode statusCode)
     {
-        public HttpStatusCode StatusCode { get; }
+        StatusCode = statusCode;
+    }
 
-        public ContainerMockException(HttpStatusCode statusCode)
-        {
-            StatusCode = statusCode;
-        }
+    public ContainerMockException(HttpStatusCode statusCode, string message)
+        : base(message)
+    {
+        StatusCode = statusCode;
+    }
 
-        public ContainerMockException(HttpStatusCode statusCode, string message)
-            : base(message)
-        {
-            StatusCode = statusCode;
-        }
+    public ContainerMockException(HttpStatusCode statusCode, string message, Exception inner)
+        : base(message, inner)
+    {
+        StatusCode = statusCode;
+    }
 
-        public ContainerMockException(HttpStatusCode statusCode, string message, Exception inner)
-            : base(message, inner)
-        {
-            StatusCode = statusCode;
-        }
+    protected ContainerMockException(
+        SerializationInfo info,
+        StreamingContext context) : base(info, context)
+    {
+    }
 
-        protected ContainerMockException(
-            SerializationInfo info,
-            StreamingContext context) : base(info, context)
-        {
-        }
-
-        public CosmosException ToCosmosException()
-        {
-            return new CosmosException(Message, StatusCode, 0, string.Empty, 0);
-        }
+    public CosmosException ToCosmosException()
+    {
+        return new CosmosException(Message, StatusCode, 0, string.Empty, 0);
     }
 }
