+++ CosmosTestHelpers/ContainerMockData/Response.cs
@@ -1,14 +1,13 @@
-﻿namespace CosmosTestHelpers.ContainerMockData
+﻿namespace CosmosTestHelpers.ContainerMockData;
+
+internal class Response
 {
-    internal class Response
+    public Response(ContainerItem item, bool isUpdate)
     {
-        public Response(ContainerItem item, bool isUpdate)
-        {
-            Item = item;
-            IsUpdate = isUpdate;
-        }
-        
-        public ContainerItem Item { get; }
-        public bool IsUpdate { get; }
+        Item = item;
+        IsUpdate = isUpdate;
     }
+        
+    public ContainerItem Item { get; }
+    public bool IsUpdate { get; }
 }
