+++ CosmosTestHelpers/CosmosTestHelpers.csproj
@@ -20,7 +20,7 @@
     </PropertyGroup>
 
      <ItemGroup>
-        <None Include="..\README.md" Pack="true" PackagePath="\"/>
+        <None Include="..\README.md" Pack="true" PackagePath="\" />
     </ItemGroup>
 
     <ItemGroup>
