+++ README.md
@@ -1,1 +1,6 @@
-# cosmos-test-helper
+# Cosmos Test Helpers
+
+A in-memory mock for a Cosmos `Container` for use in tests.
+
+> Note: This does not replicate all of a `Containers` functionality, 
+> but does cover the basics
